chore: renamed modals
