--- src/renderer/components/blocks/modals/ConfirmCreateStore/index.tsx
@@ -1,1 +1,0 @@
-export * from './ConfirmCreateStore'
