+++ src/renderer/components/blocks/modals/ConfirmCreateStoreModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './ConfirmCreateStoreModal'
