--- src/renderer/components/blocks/modals/CreateStoreError/index.tsx
@@ -1,1 +1,0 @@
-export * from './CreateStoreError';
