+++ src/renderer/components/blocks/modals/CreateStoreErrorModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './CreateStoreErrorModal';
