--- src/renderer/components/blocks/modals/CreateStoreSuccess/index.tsx
@@ -1,1 +1,0 @@
-export * from './CreateStoreSuccess'
