+++ src/renderer/components/blocks/modals/CreateStoreSuccessModal/CreateStoreSuccessModal.tsx
