+++ src/renderer/components/blocks/modals/CreateStoreSuccessModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './CreateStoreSuccessModal'
