--- src/renderer/components/blocks/modals/InvalidStoreIdError/index.tsx
@@ -1,1 +1,0 @@
-export * from './InvalidStoreIdError';
