+++ src/renderer/components/blocks/modals/InvalidStoreIdErrorModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './InvalidStoreIdErrorModal';
