--- src/renderer/components/blocks/modals/WaitingForWalletSync/index.tsx
@@ -1,1 +1,0 @@
-export * from './WaitingForWalletSync';
