+++ src/renderer/components/blocks/modals/WaitingForWalletSyncModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './WaitingForWalletSyncModal';
