--- src/renderer/components/blocks/modals/WalletBalanceInsufficient/index.tsx
@@ -1,1 +1,0 @@
-export * from './WalletBalanceInsufficient';
