+++ src/renderer/components/blocks/modals/WalletBalanceInsufficientModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './WalletBalanceInsufficientModal';
