--- src/renderer/components/blocks/modals/WalletNotSyncedError/index.tsx
@@ -1,1 +1,0 @@
-export * from './WalletNotSyncedError';
