+++ src/renderer/components/blocks/modals/WalletNotSyncedErrorModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './WalletNotSyncedErrorModal';
