+++ src/renderer/components/blocks/modals/index.tsx
@@ -1,7 +1,7 @@
-export * from './WalletNotSyncedError';
-export * from './WalletBalanceInsufficient';
-export * from './CreateStoreSuccess';
-export * from './CreateStoreError';
-export * from './WaitingForWalletSync';
-export * from './InvalidStoreIdError'
-export * from './ConfirmCreateStore'
+export * from './WalletNotSyncedErrorModal';
+export * from './WalletBalanceInsufficientModal';
+export * from './CreateStoreSuccessModal';
+export * from './CreateStoreErrorModal';
+export * from './WaitingForWalletSyncModal';
+export * from './InvalidStoreIdErrorModal'
+export * from './ConfirmCreateStoreModal'
